+++ resources/views/article/show.blade.php
@@ -15,12 +15,14 @@
     <h5 class="card-title">{{ $article->name }}</h5>
     <p class="card-text">{{ $article->desc }}</p>
     <div class="d-flex justify-content-end gap-3">
+      @can('update', $article)
         <a href="/article/{{ $article->id }}/edit" class="btn btn-primary">Edit article</a>
         <form action="/article/{{ $article->id }}" method="POST">
             @method('DELETE')
             @csrf
             <button type="submit" class="btn btn-warning">Delete article</button>
         </form>
+      @endcan
     </div>
 
     </div>
@@ -60,8 +62,10 @@
       <div class="card-body">
         <h5 class="card-title">{{$comment->name}}</h5>
         <p class="card-text">{{$comment->desc}}</p>
-        <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Edit comment</a>
-        <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Delete comment</a>
+        @can('update_comment',$comment)
+          <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Edit comment</a>
+          <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Delete comment</a>
+        @endcan
       </div>
     </div>
   </div>
